requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
pillow>=10.2.0
google-generativeai>=0.7.2
google-cloud-vision>=3.7.2
pdfplumber>=0.10.4
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
requests>=2.31.0
deep-translator>=1.11.4
